fix: splt components
