+++ ui/app/(landing-pages)/acces-recruteur/_components/AlgoRecruiter.tsx
@@ -1,52 +1,54 @@
-import { Box, Divider, Grid, GridItem, Image, ListItem, Text, UnorderedList } from "@chakra-ui/react"
+import { fr } from "@codegouvfr/react-dsfr"
+import { Box, Grid2 as Grid, List, ListItem, ListItemIcon, Typography } from "@mui/material"
+import Image from "next/image"
 
-import { DsfrLink } from "../../../../components/dsfr/DsfrLink"
+import { DsfrLink } from "@/components/dsfr/DsfrLink"
 
-export const AlgoRecruiter = ({ withLinks }) => (
-  <Box as="section" backgroundColor="white">
-    <Box py={12} backgroundColor="#f5f5fe">
-      {/* @ts-expect-error: TODO */}
-      <Grid templateColumns={{ base: "repeat(1, 1fr)", lg: "repeat(5, 1fr)" }} spacing="20px">
-        {/* @ts-expect-error: TODO */}
-        <GridItem colSpan={{ base: "1", lg: "3" }} px={{ base: 3, md: 6, lg: 12 }}>
-          <Box as="h2">
-            <Text as="span" display="block" mb={1} variant="editorialContentH1" color="#2a2a2a">
-              La bonne alternance révèle
-            </Text>
-            <Text as="span" display="block" mb={1} variant="editorialContentH1">
+export const AlgoRecruteur = ({ withLinks = false }: { withLinks?: boolean }) => {
+  return (
+    <Box sx={{ p: fr.spacing("5w"), borderRadius: "10px", backgroundColor: fr.colors.decisions.background.alt.blueFrance.default }}>
+      <Grid container spacing={fr.spacing("8w")} sx={{ alignItems: "center", justifyContent: "center" }}>
+        <Grid size={{ xs: 12, md: 7 }}>
+          <Typography sx={{ mb: 2, fontSize: "40px", fontWeight: "700", lineHeight: "48px" }}>
+            La bonne alternance révèle
+            <br />
+            <Typography component={"span"} sx={{ color: "#0063BC", fontSize: "40px", fontWeight: "700", lineHeight: "48px" }}>
               le marché caché de l&apos;emploi
-            </Text>
-          </Box>
-
-          <Divider variant="pageTitleDivider" my={12} />
-          <Box as="p" mb="3">
-            Le saviez-vous ? Afin d&apos;aider les candidats intéressés par l&apos;alternance à trouver un contrat, nous exposons différents types d&apos;entreprises sur notre
-            service :
-          </Box>
-          <UnorderedList>
-            <ListItem mb="3">
-              <strong>Celles ayant émis un besoin en recrutement </strong>sur notre plateforme ainsi que sur France Travail et ses sites partenaires
+            </Typography>{" "}
+          </Typography>
+          <Box component="hr" sx={{ maxWidth: "93px", border: "none", borderBottom: "none", borderTop: "4px solid #0063CB", opacity: 1 }} />
+          <List>
+            <ListItem sx={{ alignItems: "flex-start", px: 0 }}>
+              <ListItemIcon sx={{ minWidth: "30px" }}>
+                <Box component={"span"} className={fr.cx("ri-circle-fill", "fr-icon--xs")} />
+              </ListItemIcon>
+              <Typography>
+                <span className="fr-text--bold">Celles ayant émis un besoin en recrutement </span>sur notre plateforme ainsi que sur France Travail et ses sites partenaires
+              </Typography>
             </ListItem>
-            <ListItem mb="3">
-              <strong>Celles n&apos;ayant pas diffusé d&apos;offres, mais ayant été identifiées comme &quot;à fort potentiel d&apos;embauche en alternance&quot;</strong> grâce à
-              l'analyse de diverses données publiques (données de recrutement, données financières, etc.). La bonne alternance identifie ainsi chaque mois une liste restreinte
-              d'entreprises à fort potentiel d'embauche en alternance pour faciliter les démarches de candidatures spontanées de ses utilisateurs.
+            <ListItem sx={{ alignItems: "flex-start", px: 0 }}>
+              <ListItemIcon sx={{ minWidth: "30px" }}>
+                <Box component={"span"} className={fr.cx("ri-circle-fill", "fr-icon--xs")} />
+              </ListItemIcon>
+              <Typography>
+                <span className="fr-text--bold">
+                  Celles n&apos;ayant pas diffusé d&apos;offres, mais ayant été identifiées comme &quot;à fort potentiel d&apos;embauche en alternance&quot;
+                </span>{" "}
+                grâce à l'analyse de diverses données publiques (données de recrutement, données financières, etc.). La bonne alternance identifie ainsi chaque mois une liste
+                restreinte d'entreprises à fort potentiel d'embauche en alternance pour faciliter les démarches de candidatures spontanées de ses utilisateurs.
+              </Typography>
             </ListItem>
-          </UnorderedList>
+          </List>
           {withLinks && (
-            <DsfrLink href="/desinscription" aria-label="Accès au formulaire de désinscription au service d'envoi de candidatures spontanées">
+            <DsfrLink arrow="right" href="/desinscription" aria-label="Accès au formulaire de désinscription au service d'envoi de candidatures spontanées">
               Je ne souhaite plus recevoir de candidature spontanée
-              <Image ml={1} display="inline" src="/images/icons/arrow_right.svg" alt="" />
             </DsfrLink>
           )}
-        </GridItem>
-        {/* @ts-expect-error: TODO */}
-        <GridItem colSpan={{ base: "2" }}>
-          <Box display="flex" justifyContent="center" alignItems="center" pr={{ base: 0, lg: 12 }}>
-            <Image src="/images/icons/algo_recruiter.svg" alt="" />
-          </Box>
-        </GridItem>
+        </Grid>
+        <Grid size={{ xs: 12, md: 5 }}>
+          <Image src="/images/icons/algo_recruiter.svg" alt="" width={398} height={431} />
+        </Grid>
       </Grid>
     </Box>
-  </Box>
-)
+  )
+}
